BackendV2/Backend/WealthAdvisoryBackend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
langchain>=0.1.6
langchain-google-genai>=0.0.11
pydantic>=2.6.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
reportlab>=4.0.0
aiohttp>=3.9.0
httpx>=0.27.0
langfuse>=2.0.0
yfinance>=0.2.0
pytz>=2024.1
jinja2>=3.1.0
pdfkit>=1.0.0